requirements-devel.txt
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
netaddr==0.7.19
pluggy==0.11.0
py==1.8.0
pytest==4.5.0
pytest-cov==2.7.1
pytest-mock==1.10.4
python-iptables==0.14.0
PyYAML==5.1
requests==2.22.0
requests-mock==1.6.0
six==1.12.0
urllib3==1.25.2
vermin==0.5.0
wcwidth==0.1.7
